Avatar textrecord as a separate flow
